taskmaster_backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
